feat(dashboard): snapshot select component
