+++ apps/dashboard/src/components/Sandbox/CreateSandboxSheet.tsx
@@ -4,15 +4,11 @@
  */
 
 import { CreateResourceButton } from '@/components/CreateResourceButton'
-import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import { Checkbox } from '@/components/ui/checkbox'
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from '@/components/ui/command'
 import { Field, FieldDescription, FieldError, FieldLabel } from '@/components/ui/field'
 import { Input } from '@/components/ui/input'
-import { InputGroup, InputGroupAddon } from '@/components/ui/input-group'
 import { Label } from '@/components/ui/label'
-import { Popover, PopoverAnchor, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import {
   Sheet,
@@ -28,9 +24,7 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { useCreateSandboxMutation } from '@/hooks/mutations/useCreateSandboxMutation'
 import { useSetOrganizationDefaultRegionMutation } from '@/hooks/mutations/useSetOrganizationDefaultRegionMutation'
 import { useOrganizationUsageOverviewQuery } from '@/hooks/queries/useOrganizationUsageOverviewQuery'
-import { useSnapshotsQuery } from '@/hooks/queries/useSnapshotsQuery'
 import { useConfig } from '@/hooks/useConfig'
-import { useDebouncedValue } from '@/hooks/useDebouncedValue'
 import { useRegions } from '@/hooks/useRegions'
 import { useSelectedOrganization } from '@/hooks/useSelectedOrganization'
 import { parseEnvFile } from '@/lib/env'
@@ -38,18 +32,18 @@ import { handleApiError } from '@/lib/error-handling'
 import { GPU_TYPE_LABELS } from '@/lib/gpu-types'
 import { imageNameSchema } from '@/lib/schema'
 import { cn, getRegionFullDisplayName } from '@/lib/utils'
-import { GpuType, OrganizationUserRoleEnum, SnapshotState, type Region, type SnapshotDto } from '@daytona/api-client'
+import { GpuType, OrganizationUserRoleEnum, RegionType, type Region, type SnapshotDto } from '@daytona/api-client'
 import { Sandbox } from '@daytona/sdk'
 import { useForm, useStore } from '@tanstack/react-form'
 import { isAxiosError } from 'axios'
-import { AnimatePresence, motion } from 'framer-motion'
-import { Check, ChevronDownIcon, Info, Minus, Plus, SearchIcon, Upload } from 'lucide-react'
+import { Info, Minus, Plus, Upload } from 'lucide-react'
 import { ComponentProps, Ref, useCallback, useEffect, useImperativeHandle, useMemo, useRef, useState } from 'react'
 import { NumericFormat } from 'react-number-format'
 import { toast } from 'sonner'
 import { z } from 'zod'
 import { Tooltip } from '../Tooltip'
 import { ScrollArea } from '../ui/scroll-area'
+import { SnapshotSelect } from './SnapshotSelect'
 
 const NAME_REGEX = /^[a-zA-Z0-9][a-zA-Z0-9._-]*$/
 
@@ -173,215 +167,6 @@ const InfoTooltipButton = ({ className, ...props }: ComponentProps<'button'>) =>
   )
 }
 
-const commandInputIconMotion = {
-  initial: { opacity: 0, y: 4, scale: 0.95 },
-  animate: { opacity: 1, y: 0, scale: 1 },
-  exit: { opacity: 0, y: -4, scale: 0.95 },
-  transition: { duration: 0.14, ease: 'easeOut' },
-} as const
-
-function CommandInputStatusIcon({ fetching }: { fetching: boolean }) {
-  return (
-    <span className="relative flex size-4 shrink-0 items-center justify-center opacity-50">
-      <AnimatePresence initial={false} mode="wait">
-        {fetching ? (
-          <motion.span
-            key="fetching"
-            className="absolute inset-0 flex items-center justify-center"
-            {...commandInputIconMotion}
-          >
-            <Spinner className="size-4" />
-          </motion.span>
-        ) : (
-          <motion.span
-            key="search"
-            className="absolute inset-0 flex items-center justify-center"
-            {...commandInputIconMotion}
-          >
-            <SearchIcon className="size-4" />
-          </motion.span>
-        )}
-      </AnimatePresence>
-    </span>
-  )
-}
-
-// todo(rpavlini): pull out an async combobox component
-function SnapshotCommandSelect({
-  id,
-  value,
-  selectedSnapshot,
-  snapshots,
-  loading,
-  fetching,
-  searchValue,
-  getRegionName,
-  open,
-  popoverContainer,
-  onChange,
-  onOpenChange,
-  onSearchChange,
-}: {
-  id: string
-  value?: string
-  selectedSnapshot?: SnapshotDto
-  snapshots: SnapshotDto[]
-  loading: boolean
-  fetching: boolean
-  searchValue: string
-  getRegionName: (regionId: string) => string | undefined
-  open: boolean
-  popoverContainer?: HTMLElement | null
-  onChange: (snapshot: SnapshotDto) => void
-  onOpenChange: (open: boolean) => void
-  onSearchChange: (value: string) => void
-}) {
-  const selectedLabel = selectedSnapshot?.name ?? value ?? 'Default snapshot'
-
-  const handleChange = (snapshot: SnapshotDto) => {
-    onChange(snapshot)
-    onOpenChange(false)
-  }
-
-  return (
-    <Popover open={open} onOpenChange={onOpenChange}>
-      <PopoverAnchor asChild>
-        <InputGroup
-          className={cn('h-8 overflow-hidden', {
-            'opacity-50': loading,
-          })}
-          data-disabled={loading ? true : undefined}
-        >
-          <PopoverTrigger asChild>
-            <button
-              id={id}
-              type="button"
-              disabled={loading}
-              data-slot="input-group-control"
-              className="absolute inset-0 z-10 rounded-md outline-none disabled:cursor-not-allowed"
-              aria-label={loading ? 'Loading snapshots' : selectedLabel}
-            />
-          </PopoverTrigger>
-          <span
-            aria-hidden="true"
-            className={cn('min-w-0 flex-1 truncate px-3 text-sm', {
-              'text-muted-foreground': !value,
-            })}
-          >
-            {loading ? 'Loading snapshots...' : selectedLabel}
-          </span>
-          <InputGroupAddon align="inline-end">
-            <ChevronDownIcon aria-hidden="true" className="size-4 text-muted-foreground" />
-          </InputGroupAddon>
-        </InputGroup>
-      </PopoverAnchor>
-      <PopoverContent className="w-[var(--radix-popover-trigger-width)] p-0" align="start" container={popoverContainer}>
-        <Command shouldFilter={false}>
-          <CommandInput
-            value={searchValue}
-            onValueChange={onSearchChange}
-            placeholder="Search snapshots..."
-            icon={<CommandInputStatusIcon fetching={fetching} />}
-          />
-          <CommandList className="max-h-64 overscroll-contain">
-            {loading ? (
-              <div className="flex items-center justify-center gap-2 py-6 text-sm text-muted-foreground">
-                <Spinner className="size-4" />
-                Loading snapshots...
-              </div>
-            ) : (
-              <>
-                <CommandEmpty>{fetching ? 'Searching snapshots...' : 'No active snapshots found.'}</CommandEmpty>
-                <CommandGroup
-                  className={cn('transition-opacity duration-150 ease-out', {
-                    'opacity-50': fetching && snapshots.length > 0,
-                  })}
-                >
-                  {snapshots.map((snapshot) => (
-                    <CommandItem
-                      key={snapshot.id}
-                      value={`${snapshot.name} ${(snapshot.regionIds ?? [])
-                        .map((regionId) => getRegionName(regionId) ?? regionId)
-                        .join(' ')}`}
-                      onSelect={() => handleChange(snapshot)}
-                      className="cursor-pointer gap-2"
-                    >
-                      <Check
-                        className={cn('size-4 shrink-0', {
-                          'opacity-100': value === snapshot.name,
-                          'opacity-0': value !== snapshot.name,
-                        })}
-                      />
-                      <div className="flex min-w-0 flex-1 items-center gap-2">
-                        <span className="truncate">{snapshot.name}</span>
-                        {snapshot.general && (
-                          <Badge variant="secondary" className="shrink-0">
-                            System
-                          </Badge>
-                        )}
-                      </div>
-                      <SnapshotRegionInline regionIds={snapshot.regionIds} getRegionName={getRegionName} />
-                    </CommandItem>
-                  ))}
-                </CommandGroup>
-              </>
-            )}
-          </CommandList>
-        </Command>
-      </PopoverContent>
-    </Popover>
-  )
-}
-
-function SnapshotRegionInline({
-  regionIds,
-  getRegionName,
-}: {
-  regionIds?: string[]
-  getRegionName: (regionId: string) => string | undefined
-}) {
-  if (!regionIds?.length) {
-    return <span className="ml-auto shrink-0 text-xs text-muted-foreground">-</span>
-  }
-
-  const regionNames = regionIds.map((regionId) => getRegionName(regionId) ?? regionId)
-  const firstRegion = regionNames[0]
-  const remainingCount = regionNames.length - 1
-
-  if (remainingCount === 0) {
-    return (
-      <span
-        className="ml-auto block max-w-[150px] shrink-0 truncate text-right text-xs text-muted-foreground"
-        title={firstRegion}
-      >
-        {firstRegion}
-      </span>
-    )
-  }
-
-  return (
-    <Tooltip
-      label={
-        <div className="ml-auto flex shrink-0 items-center gap-1.5">
-          <span className="max-w-[150px] truncate text-xs text-muted-foreground">{firstRegion}</span>
-          <Badge variant="secondary" className="h-5 px-1.5 py-0 text-xs">
-            +{remainingCount}
-          </Badge>
-        </div>
-      }
-      content={
-        <div className="flex flex-col gap-1">
-          {regionNames.map((regionName, index) => (
-            <span key={`${regionName}-${index}`}>{regionName}</span>
-          ))}
-        </div>
-      }
-    />
-  )
-}
-
-const DEFAULT_SNAPSHOTS: SnapshotDto[] = []
-
 export const CreateSandboxSheet = ({
   className,
   ref,
@@ -392,17 +177,18 @@ export const CreateSandboxSheet = ({
   onSandboxCreated?: (sandbox: Sandbox) => void
 }) => {
   const [open, setOpen] = useState(false)
-  const [snapshotSelectOpen, setSnapshotSelectOpen] = useState(false)
-  const [snapshotSearchValue, setSnapshotSearchValue] = useState('')
   const [selectedSnapshotOption, setSelectedSnapshotOption] = useState<SnapshotDto | undefined>(undefined)
 
   const config = useConfig()
-  const { availableRegions: regions, loadingAvailableRegions: loadingRegions, getRegionName } = useRegions()
+  const { availableRegions: regions, loadingAvailableRegions: loadingRegions } = useRegions()
   const { selectedOrganization, authenticatedUserOrganizationMember } = useSelectedOrganization()
   const { reset: resetCreateSandboxMutation, ...createSandboxMutation } = useCreateSandboxMutation()
   const setDefaultRegionMutation = useSetOrganizationDefaultRegionMutation()
   const formRef = useRef<HTMLFormElement>(null)
-  const sheetContentRef = useRef<HTMLDivElement>(null)
+  const [popoverContainer, setPopoverContainer] = useState<HTMLDivElement | null>(null)
+  const handleSheetContentRef = useCallback((node: HTMLDivElement | null) => {
+    setPopoverContainer(node)
+  }, [])
   const formDefaultValues = useMemo<FormValues>(
     () => ({
       ...defaultValues,
@@ -427,39 +213,6 @@ export const CreateSandboxSheet = ({
   })
 
   const formSchema = useMemo(() => buildFormSchema(maxCpu, maxMemory, maxDisk), [maxCpu, maxMemory, maxDisk])
-
-  const debouncedSnapshotSearchValue = useDebouncedValue(snapshotSearchValue, 300)
-  const snapshotNameFilter = debouncedSnapshotSearchValue.trim()
-  const {
-    data: snapshotsData,
-    isLoading: snapshotsLoading,
-    isFetching: snapshotsFetching,
-  } = useSnapshotsQuery({
-    page: 1,
-    pageSize: 100,
-    filters: snapshotNameFilter ? { name: snapshotNameFilter } : undefined,
-  })
-
-  const snapshots = useMemo(
-    () => snapshotsData?.items.filter((snapshot) => snapshot.state === SnapshotState.ACTIVE) ?? DEFAULT_SNAPSHOTS,
-    [snapshotsData?.items],
-  )
-  const defaultSnapshotFromList = useMemo(
-    () => snapshots.find((snapshot) => snapshot.name === config.defaultSnapshot),
-    [config.defaultSnapshot, snapshots],
-  )
-  const shouldFetchDefaultSnapshot = Boolean(snapshotsData && !defaultSnapshotFromList)
-  const { data: defaultSnapshotData, isFetching: defaultSnapshotFetching } = useSnapshotsQuery(
-    {
-      page: 1,
-      pageSize: 10,
-      filters: { name: config.defaultSnapshot },
-    },
-    { enabled: shouldFetchDefaultSnapshot },
-  )
-
-  const snapshotsSearchPending = snapshotSearchValue !== debouncedSnapshotSearchValue
-  const snapshotsSearchActive = snapshotsFetching || snapshotsSearchPending
 
   const form = useForm({
     defaultValues: formDefaultValues,
@@ -569,24 +322,8 @@ export const CreateSandboxSheet = ({
       return undefined
     }
 
-    return snapshots.find((snapshot) => snapshot.name === selectedSnapshotName) ?? selectedSnapshotOption
-  }, [selectedSnapshotName, selectedSnapshotOption, snapshots])
-  const defaultSnapshot =
-    defaultSnapshotFromList ??
-    defaultSnapshotData?.items.find(
-      (snapshot) => snapshot.state === SnapshotState.ACTIVE && snapshot.name === config.defaultSnapshot,
-    )
-  const effectiveSnapshot = useMemo(() => {
-    if (selectedSnapshot) {
-      return selectedSnapshot
-    }
-
-    if (!selectedSnapshotName || selectedSnapshotName === config.defaultSnapshot) {
-      return defaultSnapshot
-    }
-
-    return undefined
-  }, [config.defaultSnapshot, defaultSnapshot, selectedSnapshot, selectedSnapshotName])
+    return selectedSnapshotOption?.name === selectedSnapshotName ? selectedSnapshotOption : undefined
+  }, [selectedSnapshotName, selectedSnapshotOption])
   const getRegionsForSnapshot = useCallback(
     (snapshot?: SnapshotDto) => {
       if (!snapshot) {
@@ -602,17 +339,22 @@ export const CreateSandboxSheet = ({
     },
     [regions],
   )
+  const sharedAvailableRegions = useMemo(
+    () => regions.filter((region) => region.regionType === RegionType.SHARED),
+    [regions],
+  )
   const snapshotFilteredRegions = useMemo(() => {
     if (selectedSource !== Source.SNAPSHOT) {
       return regions
     }
 
-    return getRegionsForSnapshot(effectiveSnapshot)
-  }, [effectiveSnapshot, getRegionsForSnapshot, regions, selectedSource])
-  const defaultSnapshotNeeded = !selectedSnapshotName || selectedSnapshotName === config.defaultSnapshot
-  const defaultSnapshotLoading = defaultSnapshotNeeded && !defaultSnapshot && defaultSnapshotFetching
-  const regionsDisabled =
-    loadingRegions || (selectedSource === Source.SNAPSHOT && (snapshotsSearchActive || defaultSnapshotLoading))
+    if (!selectedSnapshot) {
+      return sharedAvailableRegions
+    }
+
+    return getRegionsForSnapshot(selectedSnapshot)
+  }, [getRegionsForSnapshot, regions, selectedSnapshot, selectedSource, sharedAvailableRegions])
+  const regionsDisabled = loadingRegions
 
   const setRegionFromAvailableRegions = useCallback(
     (nextRegions: Region[]) => {
@@ -660,7 +402,6 @@ export const CreateSandboxSheet = ({
     (snapshot: SnapshotDto) => {
       form.setFieldValue('snapshot', snapshot.name)
       setSelectedSnapshotOption(snapshot)
-      setSnapshotSearchValue(snapshot.name)
       setRegionFromAvailableRegions(getRegionsForSnapshot(snapshot))
     },
     [form, getRegionsForSnapshot, setRegionFromAvailableRegions],
@@ -669,21 +410,13 @@ export const CreateSandboxSheet = ({
   const resetState = useCallback(() => {
     resetForm(formDefaultValues)
     resetCreateSandboxMutation()
-    setSnapshotSelectOpen(false)
-    setSnapshotSearchValue('')
     setSelectedSnapshotOption(undefined)
   }, [formDefaultValues, resetForm, resetCreateSandboxMutation])
 
   const handleDefaultSnapshotSelect = useCallback(() => {
-    if (defaultSnapshot) {
-      handleSnapshotChange(defaultSnapshot)
-    } else {
-      form.setFieldValue('snapshot', config.defaultSnapshot)
-      setSelectedSnapshotOption(undefined)
-      setSnapshotSearchValue(config.defaultSnapshot)
-    }
-    setSnapshotSelectOpen(false)
-  }, [config.defaultSnapshot, defaultSnapshot, form, handleSnapshotChange])
+    form.setFieldValue('snapshot', config.defaultSnapshot)
+    setSelectedSnapshotOption(undefined)
+  }, [config.defaultSnapshot, form])
 
   const handleEnvFileImport = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -734,7 +467,6 @@ export const CreateSandboxSheet = ({
     if (!open || loadingRegions) return
 
     if (selectedSource === Source.SNAPSHOT) {
-      if (snapshotsSearchActive) return
       setRegionFromAvailableRegions(snapshotFilteredRegions)
       return
     }
@@ -752,7 +484,6 @@ export const CreateSandboxSheet = ({
     selectedSource,
     setRegionFromAvailableRegions,
     snapshotFilteredRegions,
-    snapshotsSearchActive,
   ])
 
   return (
@@ -765,7 +496,7 @@ export const CreateSandboxSheet = ({
       <SheetTrigger asChild>
         <CreateResourceButton resource="Sandbox" />
       </SheetTrigger>
-      <SheetContent ref={sheetContentRef} className={`w-dvw sm:w-[500px] p-0 flex flex-col gap-0 ${className ?? ''}`}>
+      <SheetContent ref={handleSheetContentRef} className={cn('w-dvw sm:w-[500px] flex flex-col gap-0 p-0', className)}>
         <SheetHeader className="border-b border-border p-4 px-5 items-center flex text-left flex-row">
           <SheetTitle>Create Sandbox</SheetTitle>
           <SheetDescription className="sr-only">Create a new sandbox in your organization.</SheetDescription>
@@ -828,20 +559,13 @@ export const CreateSandboxSheet = ({
                       {(field) => (
                         <Field>
                           <FieldLabel htmlFor={field.name}>Snapshot</FieldLabel>
-                          <SnapshotCommandSelect
+                          <SnapshotSelect
                             id={field.name}
+                            name={field.name}
                             value={field.state.value}
-                            selectedSnapshot={selectedSnapshot}
-                            snapshots={snapshots}
-                            loading={snapshotsLoading}
-                            fetching={snapshotsSearchActive}
-                            searchValue={snapshotSearchValue}
-                            getRegionName={getRegionName}
-                            open={snapshotSelectOpen}
-                            popoverContainer={sheetContentRef.current}
-                            onChange={handleSnapshotChange}
-                            onOpenChange={setSnapshotSelectOpen}
-                            onSearchChange={setSnapshotSearchValue}
+                            placeholder="Default snapshot"
+                            popoverContainer={popoverContainer}
+                            onValueChange={handleSnapshotChange}
                           />
                           <FieldDescription>
                             If unspecified,{' '}
@@ -1086,11 +810,9 @@ export const CreateSandboxSheet = ({
                           placeholder={
                             loadingRegions
                               ? 'Loading regions...'
-                              : selectedSource === Source.SNAPSHOT && (snapshotsSearchActive || defaultSnapshotLoading)
-                                ? 'Loading snapshot regions...'
-                                : snapshotFilteredRegions.length === 0
-                                  ? 'No regions available'
-                                  : 'Select a region'
+                              : snapshotFilteredRegions.length === 0
+                                ? 'No regions available'
+                                : 'Select a region'
                           }
                         />
                       </SelectTrigger>
@@ -1103,9 +825,9 @@ export const CreateSandboxSheet = ({
                       </SelectContent>
                     </Select>
                     <FieldDescription>
-                      {selectedSource === Source.SNAPSHOT && effectiveSnapshot && snapshotFilteredRegions.length === 0
+                      {selectedSource === Source.SNAPSHOT && selectedSnapshot && snapshotFilteredRegions.length === 0
                         ? 'The selected snapshot is not available in any selectable region.'
-                        : selectedSource === Source.SNAPSHOT && effectiveSnapshot?.regionIds?.length
+                        : selectedSource === Source.SNAPSHOT && selectedSnapshot?.regionIds?.length
                           ? 'Only regions where the selected snapshot is available are shown.'
                           : 'The region where the sandbox will be created.'}
                     </FieldDescription>
